Rename "memos" module to "ephemera" to add more types
