+++ kopf/engines/probing.py
@@ -7,7 +7,7 @@ from typing import MutableMapping, Optional, Tuple
 import aiohttp.web
 
 from kopf.reactor import activities, lifecycles, registries
-from kopf.structs import callbacks, configuration, handlers, memos
+from kopf.structs import callbacks, configuration, ephemera, handlers
 
 logger = logging.getLogger(__name__)
 
@@ -20,7 +20,7 @@ _Key = Tuple[str, int]  # hostname, port
 async def health_reporter(
         endpoint: str,
         *,
-        memo: memos.AnyMemo,
+        memo: ephemera.AnyMemo,
         registry: registries.OperatorRegistry,
         settings: configuration.OperatorSettings,
         ready_flag: Optional[asyncio.Event] = None,  # used for testing
