+++ kopf/structs/callbacks.py
@@ -9,7 +9,7 @@ from typing import Any, Callable, Collection, Coroutine, NewType, Optional, Type
 
 from typing_extensions import Protocol
 
-from kopf.structs import bodies, diffs, memos, patches, primitives, references
+from kopf.structs import bodies, diffs, ephemera, patches, primitives, references
 
 # A specialised type to highlight the purpose or origin of the data of type Any,
 # to not be mixed with other arbitrary Any values, where it is indeed "any".
@@ -30,7 +30,7 @@ class ActivityFn(Protocol):
             self,
             *args: Any,
             logger: Union[logging.Logger, logging.LoggerAdapter],
-            memo: memos.AnyMemo,
+            memo: ephemera.AnyMemo,
             **kwargs: Any,
     ) -> _SyncOrAsyncResult: ...
 
@@ -51,7 +51,7 @@ class ResourceWatchingFn(Protocol):
             patch: patches.Patch,
             logger: Union[logging.Logger, logging.LoggerAdapter],
             resource: references.Resource,
-            memo: memos.AnyMemo,
+            memo: ephemera.AnyMemo,
             **kwargs: Any,
     ) -> _SyncOrAsyncResult: ...
 
@@ -73,7 +73,7 @@ class ResourceChangingFn(Protocol):
             old: Optional[Union[bodies.BodyEssence, Any]],  # "Any" is for field-handlers.
             new: Optional[Union[bodies.BodyEssence, Any]],  # "Any" is for field-handlers.
             resource: references.Resource,
-            memo: memos.AnyMemo,
+            memo: ephemera.AnyMemo,
             **kwargs: Any,
     ) -> _SyncOrAsyncResult: ...
 
@@ -92,7 +92,7 @@ class ResourceDaemonSyncFn(Protocol):
             logger: Union[logging.Logger, logging.LoggerAdapter],
             stopped: primitives.SyncDaemonStopperChecker,  # << different type
             resource: references.Resource,
-            memo: memos.AnyMemo,
+            memo: ephemera.AnyMemo,
             **kwargs: Any,
     ) -> Optional[Result]: ...
 
@@ -111,7 +111,7 @@ class ResourceDaemonAsyncFn(Protocol):
             logger: Union[logging.Logger, logging.LoggerAdapter],
             stopped: primitives.AsyncDaemonStopperChecker,  # << different type
             resource: references.Resource,
-            memo: memos.AnyMemo,
+            memo: ephemera.AnyMemo,
             **kwargs: Any,
     ) -> Optional[Result]: ...
 
@@ -132,7 +132,7 @@ class ResourceTimerFn(Protocol):
             namespace: Optional[str],
             logger: Union[logging.Logger, logging.LoggerAdapter],
             resource: references.Resource,
-            memo: memos.AnyMemo,
+            memo: ephemera.AnyMemo,
             **kwargs: Any,
     ) -> _SyncOrAsyncResult: ...
 
@@ -159,7 +159,7 @@ class WhenFilterFn(Protocol):
             old: Optional[Union[bodies.BodyEssence, Any]],  # "Any" is for field-handlers.
             new: Optional[Union[bodies.BodyEssence, Any]],  # "Any" is for field-handlers.
             resource: references.Resource,
-            memo: memos.AnyMemo,
+            memo: ephemera.AnyMemo,
             **kwargs: Any,
     ) -> bool: ...
 
@@ -179,7 +179,7 @@ class MetaFilterFn(Protocol):
             patch: patches.Patch,
             logger: Union[logging.Logger, logging.LoggerAdapter],
             resource: references.Resource,
-            memo: memos.AnyMemo,
+            memo: ephemera.AnyMemo,
             **kwargs: Any,
     ) -> bool: ...
