+++ kopf/structs/ephemera.py
