+++ tests/handling/test_activity_triggering.py
@@ -8,8 +8,8 @@ from kopf.reactor.handling import PermanentError, TemporaryError
 from kopf.reactor.lifecycles import all_at_once
 from kopf.reactor.registries import OperatorRegistry
 from kopf.storage.states import HandlerOutcome
+from kopf.structs.ephemera import Memo
 from kopf.structs.handlers import Activity, ActivityHandler, HandlerId
-from kopf.structs.memos import Memo
 
 
 def test_activity_error_exception():
