+++ tests/handling/test_cause_logging.py
@@ -9,8 +9,8 @@ import kopf
 from kopf.reactor.processing import process_resource_event
 from kopf.storage.progress import StatusProgressStorage
 from kopf.structs.containers import ResourceMemories
+from kopf.structs.ephemera import Memo
 from kopf.structs.handlers import ALL_REASONS, HANDLER_REASONS, Reason
-from kopf.structs.memos import Memo
 
 
 @pytest.mark.parametrize('cause_type', ALL_REASONS)
