+++ tests/handling/test_delays.py
@@ -11,8 +11,8 @@ from kopf.reactor.handling import TemporaryError
 from kopf.reactor.processing import process_resource_event
 from kopf.storage.states import HandlerState
 from kopf.structs.containers import ResourceMemories
+from kopf.structs.ephemera import Memo
 from kopf.structs.handlers import HANDLER_REASONS, Reason
-from kopf.structs.memos import Memo
 
 
 @pytest.mark.parametrize('cause_reason', HANDLER_REASONS)
