+++ tests/handling/test_error_handling.py
@@ -7,8 +7,8 @@ import kopf
 from kopf.reactor.handling import PermanentError, TemporaryError
 from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
+from kopf.structs.ephemera import Memo
 from kopf.structs.handlers import HANDLER_REASONS, Reason
-from kopf.structs.memos import Memo
 
 
 # The extrahandlers are needed to prevent the cycle ending and status purging.
