+++ tests/handling/test_event_handling.py
@@ -6,8 +6,8 @@ import pytest
 import kopf
 from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
+from kopf.structs.ephemera import Memo
 from kopf.structs.handlers import ALL_REASONS
-from kopf.structs.memos import Memo
 
 
 @pytest.mark.parametrize('cause_type', ALL_REASONS)
