+++ tests/handling/test_multistep.py
@@ -5,8 +5,8 @@ import pytest
 import kopf
 from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
+from kopf.structs.ephemera import Memo
 from kopf.structs.handlers import HANDLER_REASONS, Reason
-from kopf.structs.memos import Memo
 
 
 @pytest.mark.parametrize('deletion_ts', [
