+++ tests/handling/test_no_handlers.py
@@ -6,8 +6,8 @@ import pytest
 import kopf
 from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
+from kopf.structs.ephemera import Memo
 from kopf.structs.handlers import HANDLER_REASONS, ResourceChangingHandler
-from kopf.structs.memos import Memo
 
 LAST_SEEN_ANNOTATION = 'kopf.zalando.org/last-handled-configuration'
