+++ tests/handling/test_parametrization.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock
 import kopf
 from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
-from kopf.structs.memos import Memo
+from kopf.structs.ephemera import Memo
 
 
 async def test_parameter_is_passed_when_specified(resource, cause_mock, registry, settings):
