+++ tests/handling/test_retrying_limits.py
@@ -7,8 +7,8 @@ import pytest
 import kopf
 from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
+from kopf.structs.ephemera import Memo
 from kopf.structs.handlers import HANDLER_REASONS, Reason
-from kopf.structs.memos import Memo
 
 
 # The timeout is hard-coded in conftest.py:handlers().
